should the kidDetail screen be a tableview?
